Implemented RGBA8Color class
Added the RGBA8Color class which is an implementation of the RGBA8ColorFormat interface. This class allows setting 8-bit values for the red, green, blue, and alpha components of an RGBA8-scheme color. It provides methods to retrieve individual color components and convert the color to a RawColor using InternalColorImp. The class also includes a toString method for convenient debugging.